Fix underestimation of loop count

+++ run.py
@@ -112,7 +112,7 @@ def create(map_id, data_path, subsets, origin, dimensions, size):
     cmd += "rescale width={size} height={size} ! "
 
     if w != 1.0 or h != 1.0:
-        cmd += "loop number={} ! ".format(math.ceil(total / ze))
+        cmd += "loop number={} ! ".format(int(math.ceil(float(total) / ze)))
 
     cmd += "map-slice number={number} ! write minimum=0 maximum=255 jpeg-quality=85 filename={output}"
     cmd = cmd.format(**parameters)
